Correção de bugs
